Update the systemd Timer Name for Renewal
Since the same command is used to renew all the certificates installed
on a server, use "certbot-renew" for the name of the systemd service and
timer.

Appending the name of the site to be renewed to the name of the
timer means that there's a likelihood several timers, all doing the
same thing (attemp to renew all the certs on the server), will run
for each of the sites.

Signed-off-by: Jason Rogena <jasonrogena@gmail.com>