Retry creating cert if it fails
Sometimes the "Create and install cert" task fails because there's
another instance of certbot running on the host. This might happen if
certbot had already been setup on the host and it was, at the time of
running the task, running maybe to renew certs.

Retry running the task three times (spaced out witha delay of 5 seconds).

Signed-off-by: Jason Rogena <jason@rogena.me>